Updates .NET SDK version and dependencies
Upgrades the .NET SDK version to 9.0.x in the CI and API in-house feed workflows.

Also updates actions/checkout and actions/upload-artifact dependencies to their v4 versions.

Adds a step to add the in-house NuGet feed during the API in-house feed workflow. This includes configuring the feed URL, user, and API key.

Comments out the NuGet Publish step in the CI workflow.